fix(gateway): move get_meta to admin API.
