+++ tproxy/rpc/proto/tproxy_rpc.proto
@@ -91,8 +91,6 @@ service Tproxy {
   rpc RegisterCvm(RegisterCvmRequest) returns (RegisterCvmResponse) {}
   // List all ACME account URIs and the public key history of the certificates for the Content Addressable HTTPS.
   rpc AcmeInfo(google.protobuf.Empty) returns (AcmeInfoResponse) {}
-  // Summary API for inspect.
-  rpc GetMeta(google.protobuf.Empty) returns (GetMetaResponse);
 }
 
 message RenewCertResponse {
@@ -113,4 +111,6 @@ service Admin {
   rpc ReloadCert(google.protobuf.Empty) returns (google.protobuf.Empty) {}
   // Set CAA records
   rpc SetCaa(google.protobuf.Empty) returns (google.protobuf.Empty) {}
+  // Summary API for inspect.
+  rpc GetMeta(google.protobuf.Empty) returns (GetMetaResponse);
 }
