+++ tproxy/src/admin_service.rs
@@ -1,8 +1,9 @@
 use anyhow::{Context, Result};
 use ra_rpc::{CallContext, RpcCall};
+use std::time::{SystemTime, UNIX_EPOCH};
 use tproxy_rpc::{
     admin_server::{AdminRpc, AdminServer},
-    GetInfoRequest, GetInfoResponse, HostInfo, ListResponse, RenewCertResponse,
+    GetInfoRequest, GetInfoResponse, GetMetaResponse, HostInfo, ListResponse, RenewCertResponse,
 };
 
 use crate::main_service::{encode_ts, Proxy};
@@ -91,6 +92,34 @@ impl AdminRpc for AdminRpcHandler {
                 info: None,
             })
         }
+    }
+
+    async fn get_meta(self) -> Result<GetMetaResponse> {
+        let state = self.state.lock();
+        let handshakes = state.latest_handshakes(None)?;
+
+        // Total registered instances
+        let registered = state.state.instances.len();
+
+        // Get current timestamp
+        let now = SystemTime::now()
+            .duration_since(UNIX_EPOCH)
+            .context("system time before Unix epoch")?
+            .as_secs();
+
+        // Count online instances (those with handshakes in last 5 minutes)
+        let online = handshakes
+            .values()
+            .filter(|(ts, _)| {
+                // Skip instances that never connected (ts == 0)
+                *ts != 0 && (now - *ts) < 300
+            })
+            .count();
+
+        Ok(GetMetaResponse {
+            registered: registered as u32,
+            online: online as u32,
+        })
     }
 }
